Update initializers to use float32 dtype to resolve ValueError

+++ VishwamAI/scripts/model_architecture.py
@@ -12,46 +12,45 @@ class VishwamAIModel(hk.Module):
         self.tokenizer.pad_token = self.tokenizer.eos_token  # Set padding token to eos token
         self.transformer = hk.transform(
             lambda x: hk.Sequential([
-                hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
+                hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)),
                 hk.MultiHeadAttention(
                     num_heads=8,
                     key_size=64,
-                    w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")
+                    w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)
                 )(x, x, x),
-                hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
-                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
+                hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)),
+                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))
             ])(x),
             apply_rng=True
         )
         self.attention = hk.MultiHeadAttention(
             num_heads=8,
             key_size=64,
-            w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")
+            w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)
         )
         self.memory_network = hk.LSTM(128)
         self.memory_augmentation = unique_features()
-        self.dense = hk.Linear(1, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
+        self.dense = hk.Linear(1, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))
         self.advanced_features = self.add_advanced_features()
 
         # Define expert networks for Mixture of Experts (MoE) architecture
         self.num_experts = 4  # Reduced number of experts to 4
         self.experts = [hk.transform(
             lambda x: hk.Sequential([
-                hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
+                hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)),
                 hk.MultiHeadAttention(
                     num_heads=8,
                     key_size=64,
-                    w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")
+                    w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)
                 )(x, x, x),
-                hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")),
-                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
+                hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)),
+                hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))
             ])(x),
             apply_rng=True
         ) for _ in range(self.num_experts)]
 
         # Define gating mechanism
-        self.gating_network = hk.Linear(self.num_experts, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
-
+        self.gating_network = hk.Linear(self.num_experts, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))
     def __call__(self, inputs):
         if isinstance(inputs, jnp.ndarray):
             if inputs.dtype != jnp.int32:
@@ -118,20 +117,20 @@ class VishwamAIModel(hk.Module):
 
                 self.transformer_xl = hk.transform(
                     lambda x: hk.Sequential([
-                        lambda x: hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))(x),
+                        lambda x: hk.Embed(vocab_size=50257, embed_dim=512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))(x),
                         lambda x: x,
                         lambda x: hk.MultiHeadAttention(
                             num_heads=8,
                             key_size=64,
-                            w_init=hk.initializers.VarianceScaling(1.0, "fan_avg")
+                            w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32)
                         )(x, x, x),
-                        lambda x: hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))(x),
-                        lambda x: hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))(x)
+                        lambda x: hk.Linear(2048, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))(x),
+                        lambda x: hk.Linear(512, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))(x)
                     ])(x),
                     apply_rng=True
                 )
                 self.head_size = 64  # Store the head_size as an instance variable
-                self.custom_dense = hk.Linear(1, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg"))
+                self.custom_dense = hk.Linear(1, w_init=hk.initializers.VarianceScaling(1.0, "fan_avg", dtype=jnp.float32))
 
             def compute_relative_position_encoding(self, seq_length, num_heads, head_size):
                 # Ensure head_size is divisible by num_heads
